Impr - Refactoring to multiple tab files
This commit refactors the main racing-companion file into 1 for every main tab.
It will improve long term maintainability and hopefully help future development.
Directories is added for the new files as well.
